refactor: move gameUtils to core directory
