+++ src/components/Cell/Cell.tsx
@@ -1,5 +1,5 @@
 import { KeyboardEvent, MouseEvent } from "react";
-import { CellData } from "../../models/types";
+import { CellData } from "../../core/types";
 import { getCellAriaLabel } from "../../utils/accessibilityUtils";
 
 interface CellProps {
