+++ src/components/ControlPanel/ControlPanel.test.tsx
@@ -1,6 +1,6 @@
 import { cleanup, fireEvent, render, screen } from "@testing-library/react";
 import { beforeEach, expect, test, vi } from "vitest";
-import { GameStatus } from "../../models/types";
+import { GameStatus } from "../../core/types";
 import ControlPanel from "./ControlPanel";
 
 const defaultProps = {
